step_2( simple_version_refactoring )
