+++ src/main/java/org/example/PS.java
@@ -1,9 +1,19 @@
 package org.example;
 
 public class PS {
+    // main
     public static final String CYCLE = "CYCLE";
     public static final String NEXT = "NEXT";
     public static final String SETTINGS = "SETTINGS";
     public static final String ADD_CURRENT_SONG = "ADD_CURRENT_SONG";
     public static final String EXIT = "EXIT";
+
+    // song settings
+    public final static String LENGTH_CYCLE = "LENGTH_CYCLE";
+    public final static String COUNT_OF_POINTS = "COUNT_OF_POINTS";
+    public final static String MODE = "MODE";
+    public final static String BORDER_DOWN = "BORDER_DOWN";
+    public final static String MULTIPLICATION_HZ = "MULTIPLICATION_HZ";
+    public final static String VOLUME = "VOLUME";
+    public final static String SPEED = "SPEED";
 }
