+++ src/main/java/org/example/visibility/FrameSettings.java
@@ -1,10 +1,8 @@
 package org.example.visibility;
 
 import org.example.Helper;
-import org.example.Main;
-import org.example.Settings;
+import org.example.PS;
 import org.example.Song;
-import org.example.database.QueriesToDataBase;
 
 import javax.swing.*;
 import java.awt.*;
@@ -16,28 +14,32 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class FrameSettings extends JFrame{
-
     private List<JTextField> fields = new ArrayList<>();
     private boolean process = true;
-    private Song song = new Song();
+    private Song song;
 
-    public FrameSettings() {
+
+
+    public FrameSettings(Song song) {
+        /////////
+        this.song = song;
         this.addWindowListener(new WindowAdapter() {
             public void windowClosing(WindowEvent e) {
                 process = false;
             }
         });
+        ////////
 
         this.setBounds(0,0,300,400);
         this.setLayout(new FlowLayout(FlowLayout.CENTER,0,0));
 
-        this.add(createPanel(Settings.LENGTH_CYCLE,"96"));
-        this.add(createPanel(Settings.COUNT_OF_POINTS,"96"));
-        this.add(createPanel(Settings.MODE,"BORDER_DOWN"));
-        this.add(createPanel(Settings.BORDER_DOWN,"0"));
-        this.add(createPanel(Settings.MULTIPLICATION_HZ,"10"));
-        this.add(createPanel(Settings.VOLUME,"5"));
-        this.add(createPanel(Settings.SPEED,"500"));
+        this.add(createPanel(PS.LENGTH_CYCLE,"96"));
+        this.add(createPanel(PS.COUNT_OF_POINTS,"96"));
+        this.add(createPanel(PS.MODE,"DOWN"));
+        this.add(createPanel(PS.BORDER_DOWN,"0"));
+        this.add(createPanel(PS.MULTIPLICATION_HZ,"10"));
+        this.add(createPanel(PS.VOLUME,"5"));
+        this.add(createPanel(PS.SPEED,"500"));
 
         this.add(createButton("apply"));
 
@@ -75,13 +77,13 @@ public class FrameSettings extends JFrame{
         public void mouseClicked(MouseEvent e) {
             for (JTextField field : fields) {
                 switch (field.getName()){
-                    case Settings.LENGTH_CYCLE: song.setLengthCycle(Integer.parseInt(field.getText())); break;
-                    case Settings.COUNT_OF_POINTS: song.setCountOfPoints(Integer.parseInt(field.getText())); break;
-                    case Settings.MODE: song.setMode(field.getText()); break;
-                    case Settings.BORDER_DOWN: song.setBorderDown(Integer.parseInt(field.getText())); break;
-                    case Settings.MULTIPLICATION_HZ: song.setMultiplicationHz(Double.parseDouble(field.getText())); break;
-                    case Settings.VOLUME: song.setVolume(Integer.parseInt(field.getText())); break;
-                    case Settings.SPEED: song.setSpeed(Integer.parseInt(field.getText())); break;
+                    case PS.LENGTH_CYCLE: song.setLengthCycle(Integer.parseInt(field.getText())); break;
+                    case PS.COUNT_OF_POINTS: song.setCountOfPoints(Integer.parseInt(field.getText())); break;
+                    case PS.MODE: song.setMode(field.getText()); break;
+                    case PS.BORDER_DOWN: song.setBorderDown(Integer.parseInt(field.getText())); break;
+                    case PS.MULTIPLICATION_HZ: song.setMultiplicationHz(Double.parseDouble(field.getText())); break;
+                    case PS.VOLUME: song.setVolume(Integer.parseInt(field.getText())); break;
+                    case PS.SPEED: song.setSpeed(Integer.parseInt(field.getText())); break;
                 }
             }
